# Improve sorted entity adapter sorting performance