--- a/packages/toolkit/src/entities/tests/sorted_state_adapter.test.ts
+++ b/packages/toolkit/src/entities/tests/sorted_state_adapter.test.ts
@@ -654,8 +654,6 @@ describe('Sorted State Adapter', () => {
 
       console.log(
         `${name}: sortComparer called ${numSorts.toLocaleString()} times in ${duration.toLocaleString()}ms`,
-        numSorts.toLocaleString(),
-        'times',
       )
     }