This PR:

- Updates the sorting logic for sorted entity adapters to hopefully be more efficient most of the time

Previously we were always getting the list of items via `Object.values(state.entities)`. However, that is always going to give us an array of items in _insertion_ order, not the most recent _sorted_ order.

If we instead go `state.ids.map(id => state.entities[id])`, we at least start with a sorted array.

However, there's a couple tricky cases. If a set of updates replaced an item's ID, that throws things out of whack. It's easier to fall back to the old behavior.  There are also apparently a couple test cases where we can end up with duplicate IDs and have to watch for that.

On my own machine, the perf test case with 100K items gave me numbers like this with the existing code:

```
Upsert: sortComparer called: 3,058,124 times
Duration to upsert an item: 1,177 ms
Update: sortComparer called: 1,529,076 times
Duration to update an item: 1,139 ms
```

After this fix, it drops to:

```
Upsert: sortComparer called: 1,629,402 times
Duration to upsert an item: 533 ms
Update: sortComparer called: 100,032 times
Duration to update an item: 544 ms
```

That's still not _fast_, but it's a 50% improvement.

I _may_ have another idea or two around tracking updated items, removing those, and re-inserting them, but will need to try those later.