@markerikson 2024-04-19
Revised numbers if we reset the sort counter right before the upsert:

```
// Without fix
Upsert: sortComparer called: 1,529,445 times
Duration to upsert an item: 1,379 ms
Update: sortComparer called: 1,529,444 times
Duration to update an item: 1,482 ms

// With fix
Upsert: sortComparer called: 100,027 times
Duration to upsert an item: 847 ms
Update: sortComparer called: 100,032 times
Duration to update an item: 803 ms
```