@phryneas 2024-04-19
Just leaving this here - I suspect that pre-sorting the incoming array and manually merging them might save a bunch of comparisons on top (assuming that we don't only insert at the end):
```js
function merge(models: readonly T[], state: R): void {
    const entities = state.entities as Record<Id, T>
    const oldEntities = Object.values(state.ids).map((id: Id) => entities[id])
    const newEntities = models.slice().sort(sort)

    // Insert/overwrite all new/updated
    models.forEach((model) => {
      entities[selectId(model)] = model
    })

    const newSortedIds: Id[] = []
    let o = 0,
      n = 0
    while (o < oldEntities.length && n < newEntities.length) {
      const oldEntity = oldEntities[o] as T,
        oldId = selectId(oldEntity),
        newEntity = newEntities[n]

      // old entity has been overwritten by new entity, skip comparison
      if (entities[oldId] !== oldEntity) {
        o++
        continue
      }

      const comparison = sort(oldEntity, newEntity)
      if (comparison < 0) {
        newSortedIds.push(oldId)
        o++
        continue
      }
      const newId = selectId(newEntity)
      if (comparison > 0) {
        newSortedIds.push(newId)
        n++
        continue
      }
      newSortedIds.push(oldId)
      o++
      newSortedIds.push(newId)
      n++
    }
    while (o < oldEntities.length) {
      newSortedIds.push(selectId(oldEntities[o]))
      o++
    }
    while (n < newEntities.length) {
      newSortedIds.push(selectId(newEntities[n]))
      n++
    }
    if (!areArraysEqual(state.ids, newSortedIds)) {
      state.ids = newSortedIds
    }
  }
```
This was just a way around `merge` which will only work for insert/upsert, but I believe my approach would also work with updating ids - it just needs the sorted entries from before, the merged object and the inserted/updated objects (also sorted). Could just as well be a method argument like you did.
Feel free to swipe anything if you feel it's a useful approach.