@markerikson 2024-04-20
Current benchmark numbers:

**Existing logic:**

```
Original Setup: sortComparer called 1,529,087 times in 839ms
Insert One (random): sortComparer called 1,529,107 times in 1,018ms
Insert One (middle): sortComparer called 1,529,120 times in 1,311ms
Insert One (end): sortComparer called 1,529,137 times in 1,348ms
Add Many: sortComparer called 1,546,947 times in 1,283ms
Update One (end): sortComparer called 1,546,945 times in 1,338ms
Update One (middle): sortComparer called 1,546,948 times in 1,375ms
Update One (replace): sortComparer called 1,547,010 times in 1,403ms
```

**Binary Insertion / this PR:**

```
Original Setup: sortComparer called 1,529,043 times in 642ms
Insert One (random): sortComparer called 17 times in 569ms
Insert One (middle): sortComparer called 16 times in 452ms
Insert One (end): sortComparer called 17 times in 487ms
Add Many: sortComparer called 16,728 times in 519ms
Update One (end): sortComparer called 101,033 times in 590ms
Update One (middle): sortComparer called 101,035 times in 586ms
Update One (replace): sortComparer called 101,055 times in 700ms
```

I haven't tried to do any other perf profiling to measure where the _rest_ of the overhead is coming from. I'd guess that Immer is a large chunk of it.  But still, that's a lot fewer comparisons for those cases.