@markerikson 2024-04-22
Latest times with additional Immer-related optimizations:

```
Original Setup: sortComparer called 1,529,339 times in 600ms
Insert One (random): sortComparer called 16 times in 349ms
Insert One (middle): sortComparer called 16 times in 441ms
Insert One (end): sortComparer called 16 times in 377ms
Add Many: sortComparer called 16,710 times in 384ms
Update One (end): sortComparer called 101,034 times in 550ms
Update One (middle): sortComparer called 101,036 times in 492ms
Update One (replace): sortComparer called 101,055 times in 655ms
```