@unadlib 2024-05-14
I apologize for the sudden comment on this merged PR.

Here's a performance comparison test between Immer and Mutative that focuses solely on "Insert One (end)." 
The results are as follows:

```
// 10K items
Original Setup: sortComparer called 119,925 times in 40ms
Insert One (end): sortComparer called 13 times in 12ms
Use only Immer: sortComparer called 0 times in 10ms
Use only Mutative: sortComparer called 0 times in 2ms

// 100K items
Original Setup: sortComparer called 1,529,198 times in 349ms
Insert One (end): sortComparer called 17 times in 165ms
Use only Immer: sortComparer called 0 times in 120ms
Use only Mutative: sortComparer called 0 times in 24ms

// 1000K items
Original Setup: sortComparer called 18,604,782 times in 3,559ms
Insert One (end): sortComparer called 19 times in 2,859ms
Use only Immer: sortComparer called 0 times in 1,787ms
Use only Mutative: sortComparer called 0 times in 368ms
```

Here's the test code:

```ts
const initialItems = generateItems(INITIAL_ITEMS)

measureComparisons('Original Setup', () => {
  store.dispatch(entitySlice.actions.upsertMany(initialItems))
})

measureComparisons('Insert One (end)', () => {
  store.dispatch(
    entitySlice.actions.upsertOne({
      id: nanoid(),
      position: 0.9998,
      name: 'test',
    }),
  )
})

measureComparisons('Use only Immer', () => {
  produce(store.getState(), (draft) => {
    const id = nanoid();
    draft.entity.ids.push(id)
    draft.entity.entities[id] = {
      id,
      position: 0.9998,
      name: 'test',
    }
  })
})

measureComparisons('Use only Mutative', () => {
  create(store.getState(), (draft) => {
    const id = nanoid();
    draft.entity.ids.push(id)
    draft.entity.entities[id] = {
      id,
      position: 0.9998,
      name: 'test',
    }
  })
})
```

I know this performance test is quite specific. If you have any further ideas or questions, I’d be more than happy to discuss them.